Se agrega funcionalidad de operar con decimales al decir "CON" o "COMA"

+++ app/src/main/java/com/example/vozoperadora/Interfaz.java
@@ -5,7 +5,7 @@ import android.widget.TextView;
 public class Interfaz {
 
 
-    public void mostrarResultado(int resultado, TextView tv){
+    public void mostrarResultado(double resultado, TextView tv){
         String texto = String.valueOf(resultado);
         tv.setText(texto);
     }
